homework 7
