--- Kukurudza/src/course_1/lecture/addWorks/addWServ/AddWSr.java
@@ -1,16 +1,0 @@
-package course_1.lecture.addWorks.addWServ;
-
-import course_1.lecture.addWorks.addWorks.AddW1;
-
-
-public class AddWSr {
-
-    public void addW() {
-
-        AddW1 addW1 = new AddW1();
-        int num = addW1.NUM;
-        System.out.println("Add works №" + num);
-
-    }
-}
-
