--- Kukurudza/src/course_1/lecture/homeWork/HW.java
@@ -1,8 +1,0 @@
-package course_1.lecture.homeWork;
-
-public class HW {
-
-    private int id = 1;
-    int hW = 1;
-
-}
