--- Kukurudza/src/course_1/lecture/homeWork/HWServis.java
@@ -1,12 +1,0 @@
-package course_1.lecture.homeWork;
-
-public class HWServis {
-
-    public void hWServ() {
-
-        HW hw = new HW();
-        int nuM = hw.hW;
-        System.out.println("Home works №" + nuM);
-
-    }
-}
