--- Kukurudza/src/course_1/lecture/lec/lec/Lec1.java
@@ -1,10 +1,0 @@
-package course_1.lecture.lec.lec;
-
-public class Lec1 {
-
-    public int id;
-
-    public Lec1(int id) {
-        this.id = id;
-    }
-}
