--- Kukurudza/src/course_1/lecture/lec/lec/Lec3.java
@@ -1,14 +1,0 @@
-package course_1.lecture.lec.lec;
-
-public class Lec3 {
-
-    public int id;
-
-    public Lec3(int id) {
-        this.id = id;
-    }
-}
-
-
-
-
