--- Kukurudza/src/course_1/lecture/lec/lec/Lec4.java
@@ -1,10 +1,0 @@
-package course_1.lecture.lec.lec;
-
-public class Lec4 {
-
-    public int id;
-
-    public Lec4(int id) {
-        this.id = id;
-    }
-}
