--- Kukurudza/src/course_1/lecture/lec/lec/Lec5.java
@@ -1,11 +1,0 @@
-package course_1.lecture.lec.lec;
-
-public class Lec5 {
-
-    public int id;
-
-    public Lec5(int id) {
-        this.id = id;
-    }
-}
-
