--- Kukurudza/src/course_1/lecture/lec/lec/Lec6.java
@@ -1,10 +1,0 @@
-package course_1.lecture.lec.lec;
-
-public class Lec6 {
-
-    public int id;
-
-    public Lec6(int id) {
-        this.id = id;
-    }
-}
