--- Kukurudza/src/course_1/student/student/St2.java
@@ -1,8 +1,0 @@
-package course_1.student.student;
-
-public class St2 {
-
-    private int id = 2;
-    public final static int ST2 = 2; //student number
-
-}
