--- Kukurudza/src/course_1/teacher/teacher/Tch1.java
@@ -1,8 +1,0 @@
-package course_1.teacher.teacher;
-
-public class Tch1 {
-
-    private int id=1;
-    public final static int NUM = 1; //teacher's number
-
-}
