--- Kukurudza/src/course_1/teacher/teacher/Tch2.java
@@ -1,8 +1,0 @@
-package course_1.teacher.teacher;
-
-public class Tch2 {
-
-    private int id=2;
-    public final static int NUM = 2; //teacher's number
-
-}
