+++ Kukurudza/src/homeWork7/course_1/Course.java
@@ -1,0 +1,11 @@
+package homeWork7.course_1;
+
+public class Course {
+    public int id = 1;
+
+    public Course(int id) {
+        this.id = id;
+    }
+}
+
+
