+++ Kukurudza/src/homeWork7/course_1/lecture/addWorks/addWorks/AddW1.java
@@ -1,0 +1,10 @@
+package homeWork7.course_1.lecture.addWorks.addWorks;
+
+public class AddW1 {
+
+    private int id = 1;
+    public final int NUM = 1; //number of additional work
+
+}
+
+
