+++ Kukurudza/src/homeWork7/course_1/lecture/homeWork/HW.java
@@ -1,0 +1,8 @@
+package homeWork7.course_1.lecture.homeWork;
+
+public class HW {
+
+    private int id = 1;
+    int hW = 1;
+
+}
