+++ Kukurudza/src/homeWork7/course_1/lecture/homeWork/HWServis.java
@@ -1,0 +1,12 @@
+package homeWork7.course_1.lecture.homeWork;
+
+public class HWServis {
+
+    public void hWServ() {
+
+        HW hw = new HW();
+        int nuM = hw.hW;
+        System.out.println("Home works №" + nuM);
+
+    }
+}
