+++ Kukurudza/src/homeWork7/course_1/lecture/lec/lec/Lec2.java
@@ -1,0 +1,10 @@
+package homeWork7.course_1.lecture.lec.lec;
+
+public class Lec2 {
+
+    public int id;
+
+    public Lec2(int id) {
+        this.id = id;
+    }
+}
