+++ Kukurudza/src/homeWork7/course_1/lecture/lec/lec/Lec3.java
@@ -1,0 +1,14 @@
+package homeWork7.course_1.lecture.lec.lec;
+
+public class Lec3 {
+
+    public int id;
+
+    public Lec3(int id) {
+        this.id = id;
+    }
+}
+
+
+
+
