+++ Kukurudza/src/homeWork7/course_1/lecture/lec/lec/Lec4.java
@@ -1,0 +1,10 @@
+package homeWork7.course_1.lecture.lec.lec;
+
+public class Lec4 {
+
+    public int id;
+
+    public Lec4(int id) {
+        this.id = id;
+    }
+}
