+++ Kukurudza/src/homeWork7/course_1/lecture/lec/lec/Lec5.java
@@ -1,0 +1,11 @@
+package homeWork7.course_1.lecture.lec.lec;
+
+public class Lec5 {
+
+    public int id;
+
+    public Lec5(int id) {
+        this.id = id;
+    }
+}
+
