+++ Kukurudza/src/homeWork7/course_1/lecture/lec/lec/Lec6.java
@@ -1,0 +1,10 @@
+package homeWork7.course_1.lecture.lec.lec;
+
+public class Lec6 {
+
+    public int id;
+
+    public Lec6(int id) {
+        this.id = id;
+    }
+}
