+++ Kukurudza/src/homeWork7/course_1/student/student/St1.java
@@ -1,0 +1,8 @@
+package homeWork7.course_1.student.student;
+
+public class St1 {
+
+    private int id = 1;
+    public final static int ST1 = 1; //student number
+
+}
