+++ Kukurudza/src/homeWork7/course_1/student/student/St2.java
@@ -1,0 +1,11 @@
+package homeWork7.course_1.student.student;
+
+public class St2 {
+
+    public int id = 2;
+    public final static int ST2 = 2; //student number
+
+    public St2(int id) {
+        this.id = id;
+    }
+}
