+++ Kukurudza/src/homeWork7/course_1/teacher/teacher/Tch1.java
@@ -1,0 +1,8 @@
+package homeWork7.course_1.teacher.teacher;
+
+public class Tch1 {
+
+    private int id=1;
+    public final static int NUM = 1; //teacher's number
+
+}
