+++ Kukurudza/src/homeWork7/course_1/teacher/teacher/Tch2.java
@@ -1,0 +1,12 @@
+package homeWork7.course_1.teacher.teacher;
+
+public class Tch2 {
+
+    public int id = 2;
+    public final static int NUM = 2; //teacher's number
+
+    public Tch2(int id) {
+        this.id = id;
+    }
+}
+
